Add some actual modules to stdlib
