+++ compiler/src/Language/Mimsa/Actions/Helpers/Parse.hs
@@ -1,4 +1,4 @@
-module Language.Mimsa.Actions.Helpers.Parse (parseExpr, parseDataType) where
+module Language.Mimsa.Actions.Helpers.Parse (parseExpr, parseDataType, parseModule) where
 
 import Control.Monad.Except
 import Data.Text (Text)
@@ -7,6 +7,8 @@ import qualified Language.Mimsa.Parser as Parser
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Identifiers
+import qualified Language.Mimsa.Modules.Parse as Module
+import Language.Mimsa.Types.Modules
 
 parseExpr :: Text -> ActionM (Expr Name Annotation)
 parseExpr
@@ -19,3 +21,7 @@ parseDataType input =
   case Parser.parseTypeDecl input of
     Right a -> pure a
     Left e -> throwError (ParseError input e)
+
+parseModule :: Text -> ActionM (Module Annotation)
+parseModule = 
+  liftEither . Module.parseModule
