+++ compiler/static/modules/Maybe.mimsa
@@ -1,0 +1,11 @@
+export type Maybe a = Just a | Nothing
+
+export def fromMaybe (val: Maybe a) (fallback: a): a =
+  match val with
+    Just a -> a
+    | _ -> fallback
+
+export def fmap (f: a -> b) (maybeA: Maybe a): Maybe b
+  = match maybeA with
+      Just a -> Just (f a)
+      | _ -> Nothing
