+++ compiler/static/modules/Prelude.mimsa
@@ -1,0 +1,16 @@
+export def id a = a
+
+export def compose f g a = f (g a)
+
+export def not a = if a then False else True
+
+export def and a b = if a then b else False
+
+export def or a b = if a then True else b
+  
+export def fst pair = let (a,_) = pair in a
+
+export def snd pair = let (_,b) = pair in b
+
+export def const a b = a
+
